[inductor] realize_into should not alias src and dst
Fixes #107995

In the reproducer we have the fx graph:
```python
class <lambda>(torch.nn.Module):
    def forward(self, arg0_1: f32[1]):
        # File: <ipython-input-1-5f62cb746ad5>:10, code: return self.layer1(inputs)
        gt: b8[1] = torch.ops.aten.gt.Scalar(arg0_1, 0)
        mul: f32[1] = torch.ops.aten.mul.Tensor(arg0_1, 5.2955089)
        where: f32[1] = torch.ops.aten.where.self(gt, arg0_1, mul);  gt = mul = None

        # No stacktrace found for following nodes
        copy_: f32[1] = torch.ops.aten.copy_.default(arg0_1, where);  arg0_1 = None
        return (where,)
```

The `where` node is both copied into `arg0_1` and returned as the output of the
function. Currently `realize_into` converts the where's storage into a
`MutationLayout` of `arg0_1`, for which no tensor named `buf0` is allocated.

This is incorrect as `where` and `arg0_1` shouldn't share storage. It also
breaks the wrapper code generation which references `buf0` directly in the
return, but never allocates a `buf0`.

This issue only appears for size zero tensors, because otherwise the `src`
buffer becomes a user of `arg0_1` which forces this copy to happen anyway.

ghstack-source-id: 89b5193a21cd55a61801fedf5d92a6979dd5b049
Pull Request resolved: https://github.com/pytorch/pytorch/pull/108126